Remove `SpeechRegion` and related utilities from `support::regions` and relocate functionality to `voice_activity_filter`. Update imports across the Whisper kit to reflect this reorganization.
