+++ crates/public/bunsen/src/kits/speech/whisper/driver/support/mod.rs
@@ -2,12 +2,9 @@
 //! filter produces and the segment splitting of a decoded window. Nothing
 //! here is part of the driver's API.
 
-mod regions;
 mod segments;
 mod util;
 
-#[doc(inline)]
-pub use regions::*;
 #[doc(inline)]
 pub use segments::*;
 #[doc(inline)]
